#106 Added new methods to retrieve pending and rejected connections
